requirements.txt
fastapi
uvicorn[standard]
tensorflow-lite
scikit-learn
numpy
pandas
joblib